Refactor backend services: Fix compilation, update dependencies, and enhance consistency

+++ backend-refactor/message-service/src/main/java/com/wework/platform/message/service/impl/MessageServiceImpl.java
@@ -107,7 +107,7 @@ public class MessageServiceImpl implements MessageService {
                 .map(this::convertToDTO)
                 .collect(Collectors.toList());
 
-        return PageResult.of(dtoList, result.getTotal(), pageNum, pageSize);
+        return PageResult.of(dtoList, result.getTotal(), pageNum.longValue(), pageSize.longValue());
     }
 
     @Override
